% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{AIC.SEM}
\alias{AIC.SEM}
\title{AIC.SEM computes AIC for SEM Model}
\usage{
\method{AIC}{SEM}(subdata, SEM = SEM)
}
\arguments{
\item{SEM}{the SEM version}

\item{data}{the longline data}
}
\value{
the AIC criterion
}
\description{
AIC.SEM computes AIC for SEM Model
}

