% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlineData.R
\name{ExtractLongline}
\alias{ExtractLongline}
\title{ExtractLongline is a longline object extracts from the argument, which matches the condition fact1 and fact2}
\usage{
ExtractLongline(x, fact1 = NULL, fact2 = NULL, ind = NULL)
}
\arguments{
\item{x}{a object of class longline}

\item{fact1}{the value of Fact1 to be used for extraction}

\item{fact2}{the value of Fact2 to be used for extraction}

\item{ind}{if provided, only the individuals ind will be extracted whatever fact2 and  fact1 are}
}
\value{
a longline object with only observations matching Fact1=fact1 and/or  optionnaly Fact2=fact2
}
\description{
ExtractLongline is a longline object extracts from the argument, which matches the condition fact1 and fact2
}
\examples{
N <- 20
n <- rep(200,N)
s <- rep(100, N)
dataSim <- rmultinom(n=N, size=n, prob=c(.3, .1, .2, .4) )
testLongline <- longline(fact1 = as.factor(sample(2004:2006, size=N,
replace=TRUE)), nb=dataSim[1,], nt=dataSim[2,], nnt=dataSim[3,], ne=dataSim[4,], s=s)
ex <- ExtractLongline(testLongline, fact1="2004")
print(ex)
}

