% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatData.R
\name{FormatDataForJags}
\alias{FormatDataForJags}
\title{FormatDataForJags prepare the data to be used with MEM model in jags}
\usage{
FormatDataForJags(llData, parPrior = c(1, 1), tau.lambda = 0.5)
}
\arguments{
\item{llData}{an object of class longline}

\item{parPrior}{a 2 components vector on the probability of espace pe~dbeta(parPrior[1], parPrior[2])}

\item{tau.lambda}{the precision used in the lognormal prior for all log.mu, log.lambda*F* parameters}
}
\value{
a list to be used as data.list in jags.model with model MEM
}
\description{
FormatDataForJags prepare the data to be used with MEM model in jags
}

