% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatData.R
\name{FormatInitForJags}
\alias{FormatInitForJags}
\title{FormatInitForJags prepare the initial values to be used with MEM model in jags}
\usage{
FormatInitForJags(llData, peConst = F, MEM = 1)
}
\arguments{
\item{llData}{an object of class longline}

\item{peConst}{should the probability of escape be considered as constant over Fact1, default is FALSE}

\item{MEM}{stands for the version of MEM 1, if escape only arises from non target species, and 2 if all species have the same probability of escape}
}
\value{
a list to be used as data.list in jags.model with model MEM
}
\description{
FormatInitForJags prepare the initial values to be used with MEM model in jags
}

