% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{LogLike.SEM2}
\alias{LogLike.SEM2}
\title{LogLike.SEM1 the log likelihood given data for the parameter theta}
\usage{
LogLike.SEM2(theta.sem2, subdata)
}
\arguments{
\item{theta.sem2}{parameter relevant for SEM 2 model in log scale}

\item{theta.sem1}{parameter relevant for SEM 1 model in log scale}

\item{data}{a object of class longline}

\item{data}{a object of class longline}
}
\value{
the log likelihood for SEM1
LogLike.SEM2 the log likelihood given data for the parameter theta

the log likelihood for SEM2
}
\description{
LogLike.SEM1 the log likelihood given data for the parameter theta
}

