% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{LogLike.SEM3}
\alias{LogLike.SEM3}
\title{LogLike.SEM3 the log likelihood given data for the parameter theta}
\usage{
LogLike.SEM3(theta.sem3, subdata)
}
\arguments{
\item{theta.sem3}{parameter relevant for SEM 3 model in log scale}

\item{data}{a object of class longline}
}
\value{
the log likelihood for SEM3
}
\description{
LogLike.SEM3 the log likelihood given data for the parameter theta
}

