% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{long_line_example}
\alias{long_line_example}
\title{Exposure to Russian novels.}
\format{A data frame with 96 rows and 6 variables:
\describe{
  \item{subject}{subject names, character}
  \item{exposure}{exposure types, character}
  \item{enroll}{enrollment time, numeric}
  \item{exit}{exit from study, integer}
  \item{exposure_start}{start of exposure, integaer}
  \item{exposure_end}{end of exposure, integer}
}}
\usage{
long_line_example
}
\description{
A data frame demonstrating the format of input into the \code{\link{longlineplot}} function.
}
\keyword{datasets}

