% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/longlineData.R
\name{longline}
\alias{longline}
\title{longline is used to create a longline object.}
\usage{
longline(fact1, fact2 = NULL, nt, nnt, nb, ne = NULL, s)
}
\arguments{
\item{fact1}{a factor vector, for which which indices will be computed}

\item{fact2}{an opional  factor vector which should be accounted for in the
definition of the indices}

\item{nt}{a numeric vector of the number of individuals of target species
caught}

\item{nnt}{a numeric vector of the number of individuals of non target
species caught}

\item{nb}{a numeric vector of  the number of still baited hooks}

\item{ne}{an optional vector of  the number of empty hooks, if NULL,
then empty hooks are considered as missing information}

\item{s}{a numeric vector of soaktime}
}
\value{
an object of class longline
}
\description{
longline is used to create a longline object.
}
\examples{
N <- 20
n <- rep(200,N)
s <- rep(100, N)
dataSim <- rmultinom(n=N, size=n, prob=c(.3, .1, .2, .4) )
testLongline <- longline(fact1 = as.factor(sample(2004:2006, size=N,
replace=TRUE)),  nb=dataSim[1,], nt=dataSim[,2], nnt=dataSim[,3], ne=dataSim[,4], s=s)
}

