% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlineEx.R
\docType{data}
\name{longlineEx}
\alias{longlineEx}
\title{Example for longline data}
\usage{
longlineEx
}
\description{
This data set shows an example of a longline experiment
}
\author{
Marie-Pierre Etienne
}

