% Generated by roxygen2 (4.0.1): do not edit by hand
\name{longlineplot}
\alias{longlineplot}
\title{Plot longitudinal exposure data.}
\usage{
longlineplot(long_line_df = long_line_example, rect_fill = "grey80",
  rect_color = NA, text_size = 12, x_label = "Time (Days)",
  y_label = "")
}
\arguments{
\item{long_line_df}{a data frame formated according to long_line_example}

\item{rect_fill}{color of fill for bar denoting enrollment period}

\item{rect_color}{color of line for bar denoting enrollment period}

\item{text_size}{point size of text for axis and subject labels}

\item{x_label}{label for x-axis}

\item{y_label}{label for y-axis}
}
\value{
a ggplot plot object
}
\description{
Generates a plot with horizontal bar for enrollment period and horizontal lines for exposure duration.
}
\examples{
longlineplot(long_line_example,"grey40",NA,10,"","Subjects")
}
\seealso{
\code{\link{ggplot}}, \code{\link{geom_segment}}
}

