% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM.R
\name{MEM.MLE}
\alias{MEM.MLE}
\title{MEM.MLE computes MLE for data with several years and area, one lambda1, lambda2, pe computes for each(year,area)}
\usage{
MEM.MLE(llData, MEM = 1, sameS = T)
}
\arguments{
\item{llData}{a object of class longline}

\item{MEM}{the type of model (possible values 1 or 2)}

\item{sameS, }{should all soaktime be considered as equal ?}
}
\value{
a vetor estimates of hat.lambdaT, hat.lambda.NT, hat.pT, hat.pNT
}
\description{
MEM.MLE computes MLE for data with several years and area, one lambda1, lambda2, pe computes for each(year,area)
}

