% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graph.R
\name{PostIndicesPlot}
\alias{PostIndicesPlot}
\title{PostIndicesPlot is designed to produce plots for the posterior distribution of abundance indices for Target and non Target species, it may be used to produce posteriori boxplot of other quantitties}
\usage{
PostIndicesPlot(llData, outSamples, who = "TF1", fileOut = "plot.pdf",
  yLim = NULL, xLab = NULL)
}
\arguments{
\item{llData}{an object of class longline}

\item{outSamples}{the results of RunBayesEstimation on llData}

\item{who}{the quantity to be plotted if TF1 (resp NTF1 ) it produces the posterior distribution for target (resp Non Targe) species for every level of factor F1, Who may be any other posterior quantity of the model (log.mu1, log.mu2, log.lambda1F1, log.lambda2.F1, pe)}

\item{fileOut}{the file to save the plots}

\item{yLim}{optionnal the limits of the Y axis, default is null and the scales will be determined automatically}

\item{xLab}{optionnal the names of xlabel}
}
\value{
no return value
}
\description{
PostIndicesPlot is designed to produce plots for the posterior distribution of abundance indices for Target and non Target species, it may be used to produce posteriori boxplot of other quantitties
}

