% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LonglineBayes.R
\name{RunBayesEstimation}
\alias{RunBayesEstimation}
\title{RunBayesEstimation runs the estimation of lambdaT and lambdaNT and produce abundance indices}
\usage{
RunBayesEstimation(llData, MEM = 1, nIter = 2000, burnin = 500,
  filePath = "", parPrior = c(1, 1), peConst = F, tau.lambda = 0.5)
}
\arguments{
\item{llData}{an object of class longline}

\item{MEM}{the version of the model, default is 1}

\item{nIter}{number of MCMC iterations kept, default is 2000}

\item{burnin}{number of MCMC iterations dropped for the bruning period}

\item{filePath}{the path where results and figure should be stored}

\item{parPrior}{a 2 components vector on the probability of espace pe~dbeta(parPrior[1], parPrior[2])}

\item{peConst}{should the probability of escape be considered as constant, default is FALSE prior for all log.mu*, log.lambda*F* parameters, default value=0.5}

\item{tau.lambda}{the precision used in the lognormal}
}
\value{
a list of abundance indice, and write some figures and tables in the directory specified in pathFile
}
\description{
RunBayesEstimation runs the estimation of lambdaT and lambdaNT and produce abundance indices
}

