% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{SEM.MLE.YA}
\alias{SEM.MLE.YA}
\title{SEM.MLE.YA computes MLE for data for one specific year, one lambda1, lambda2, computed for each(area)}
\usage{
SEM.MLE.YA(subdata, sameN = TRUE, sameS = TRUE, SameSigma = 1,
  verbose = F, SEM = SEM)
}
\arguments{
\item{sameN, }{should all initial number of baits be considered as equal ?}

\item{sameS, }{should all soaktime be considered as equal ?}

\item{verbose, }{for debugging purpose}

\item{SEM, }{the version of the SEM model to be calculated : possible value 1 (if NE and NNT are grouped together, same variance), 2 (NE and NNT split into 2 categories, Same variance), 3(NE and NNT together, two different variance sigmaT and SigmaNT)}

\item{llData}{a object of class longline}

\item{sameSigma, }{should the variance be chosen equal for target and Non Target}
}
\value{
a vetor estimates of hat.lambdaT, hat.lambda.NT,  hat.lambdaNe, hat.sigmaT, hat.sigmaNT
}
\description{
SEM.MLE.YA computes MLE for data for one specific year, one lambda1, lambda2, computed for each(area)
}

