% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{Sem2MleGeneral}
\alias{Sem2MleGeneral}
\title{Sem2MleSameS computes MLE for data for one specific year, and one specific area}
\usage{
Sem2MleGeneral(subdata, verbose = F)
}
\arguments{
\item{verbose}{for debugging purpose}

\item{data}{a object of class longline}
}
\value{
a vetor estimates of hat.lambdaT, hat.lambda.NT,  hat.lambdaNe, hat.sigmaT, hat.sigmaNT
}
\description{
Sem2MleSameS computes MLE for data for one specific year, and one specific area
}

