% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{Sem2MleSameS}
\alias{Sem2MleSameS}
\title{Sem2MleSameS computes MLE for data for one specific year, and one specific area when all longlinesets share the same soaktime
if the soaktime are different a mean soaktime is computed}
\usage{
Sem2MleSameS(data)
}
\arguments{
\item{data}{a object of class longline}
}
\value{
a vetor estimates of hat.lambdaT, hat.lambda.NT,  hat.lambdaNe, hat.sigmaT, hat.sigmaNT
}
\description{
Sem2MleSameS computes MLE for data for one specific year, and one specific area when all longlinesets share the same soaktime
if the soaktime are different a mean soaktime is computed
}

