% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/longlineData.R
\name{ComputeCPUE.longline}
\alias{ComputeCPUE.longline}
\title{ComputeCPUE.longline computes classical CPUE according to Fact1 and Fact2}
\usage{
ComputeCPUE.longline(x)
}
\arguments{
\item{x}{a longline object}
}
\value{
classical CPUE  computed for every level of Fact1 and Fact2
}
\description{
ComputeCPUE.longline computes classical CPUE according to Fact1 and Fact2
}
\examples{
data(longlineEx)
ll1 <- longline(fact1=longlineData$Year, fact2=longlineData$Area,  nt=longlineData$NT, nnt=longlineData$NNT, nb=longlineData$Nb, ne=longlineData$Ne, s=longlineData$soaktime)
ComputeCPUE.longline(ll1)
ll2 <- longline(fact1=longlineData$Year, nt=longlineData$NT, nnt=longlineData$NNT, nb=longlineData$Nb, ne=longlineData$Ne, s=longlineData$soaktime)
ComputeCPUE.longline(ll2)
}

